Fixes to frontend stuff
