+++ caiman-test/src/pipelines.ron
